finance_update 2 copy/requirements.txt
# Core dependencies
python-dotenv>=1.0.0
asyncio>=3.4.3
aiohttp>=3.8.1
backoff>=2.2.1

# API clients
openai>=1.3.0
dropbox>=11.36.0

# Data processing
nltk>=3.8.1
tiktoken>=0.5.1
PyPDF2>=3.0.0
pdfminer.six>=20221105  # For PDF text extraction

# Email handling
secure-smtplib>=0.1.1

# Logging and monitoring
structlog>=23.1.0
prometheus_client>=0.19.0

# Testing and development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
